requirements.txt
pandas~=2.3.3
sqlalchemy~=2.0.44
psycopg2-binary==2.9.9
python-dotenv==1.0.0
dbt-postgres
streamlit==1.31.0
plotly==5.18.0